Declare each component in its own file
