+++ HW7/db.json
@@ -1,0 +1,22 @@
+{
+  "tasks": [
+    {
+      "title": "Task 1",
+      "isCompleted": false,
+      "id": 1,
+      "tag": "other"
+    },
+    {
+      "id": 2,
+      "title": "Task 2",
+      "isCompleted": false,
+      "tag": "work"
+    },
+    {
+      "id": 3,
+      "title": "Task 3",
+      "isCompleted": false,
+      "tag": "health"
+    }
+  ]
+}
