+++ HW7/src/components/Button.js
@@ -1,0 +1,6 @@
+export function Button({text, onClick}) {
+    const button = document.createElement("button");
+    button.innerHTML = text;
+    button.onclick = onClick;
+    return button;
+}
