+++ HW7/src/components/SearchInput.js
@@ -1,0 +1,8 @@
+export function SearchInput({searchString, placeholder, changeSearchString}) {
+    const searchInput = document.createElement("INPUT");
+    searchInput.setAttribute("type", "text");
+    searchInput.placeholder = placeholder;
+    searchInput.value = searchString;
+    searchInput.addEventListener("input", () => changeSearchString());
+    return searchInput;
+}
