+++ HW7/src/components/TextInput.js
@@ -1,0 +1,7 @@
+export function TextInput({placeholder, onInput }) {
+    const textInput = document.createElement("INPUT");
+    textInput.setAttribute("type", "text");
+    textInput.placeholder = placeholder;
+    textInput.oninput = onInput;
+    return textInput;
+}
