+++ HW7/src/useSearchString.js
@@ -1,0 +1,9 @@
+export function useSearchString(initialValue) {
+    searchString = searchString || initialValue;
+
+    function setValue(newValue) {
+        searchString = newValue;
+        renderApp();
+    }
+    return [searchString, setValue];
+}
